+++ src/Knight.cpp
@@ -1,6 +1,6 @@
 #include "Knight.h"
 
 
-Knight::Knight(int x, int y, bool colour) : Piece(x,y, colour) {
+Knight::Knight(unsigned short x, unsigned short y, bool colour) : Piece(x,y, colour) {
     figure = 'k';
 }
