+++ src/Knight.h
@@ -6,7 +6,7 @@ class Knight : public Piece
 {
 
 public:
-    Knight(int x, int y, bool colour);
+    Knight(unsigned short x, unsigned short y, bool colour);
 };
 
 #endif
